@@ -1,5 +1,4 @@
-﻿using System;
-using System.Windows.Forms;
+﻿using System.Runtime.Versioning;
 
 namespace StationMonitor
 {
@@ -9,6 +8,7 @@ namespace StationMonitor
         /// The main entry point for the application.
         /// </summary>
         [STAThread]
+        [SupportedOSPlatform("windows")]
         static void Main()
         {
             Application.EnableVisualStyles();
