@@ -6,18 +6,26 @@ using ExpertElectronics.Tci.Interfaces;
 
 namespace ExpertElectronics.Tci.TciCommands
 {
+    /// <summary>
+    /// Represents a Cw Macro speed down command for the TCI device.
+    /// </summary>
     public class TciCwMacroSpeedDownCommand : ITciCommand, IDisposable
     {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="TciCwMacroSpeedDownCommand"/> class.
+        /// </summary>
+        /// <param name="transceiverController">The transceiver controller used to update CW macro speed down state.</param>
+        private TciCwMacroSpeedDownCommand(ITransceiverController transceiverController)
+        {
+            _transceiverController = transceiverController;
+        }
         public static TciCwMacroSpeedDownCommand Create(ITransceiverController transceiverController)
         {
             Debug.Assert(transceiverController != null);
             return new TciCwMacroSpeedDownCommand(transceiverController);
         }
 
-        private TciCwMacroSpeedDownCommand(ITransceiverController transceiverController)
-        {
-            _transceiverController = transceiverController;
-        }
+
 
         public static string Name => "cw_macros_speed_down";
 
