@@ -6,18 +6,26 @@ using ExpertElectronics.Tci.Interfaces;
 
 namespace ExpertElectronics.Tci.TciCommands
 {
+    /// <summary>
+    /// Represents a Cw Macro speed up command for the TCI device.
+    /// </summary>
     public class TciCwMacroSpeedUpCommand : ITciCommand, IDisposable
     {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="TciCwMacroSpeedUpCommand"/> class.
+        /// </summary>
+        /// <param name="transceiverController">The transceiver controller used to update CW macro speed up state.</param>
+        private TciCwMacroSpeedUpCommand(ITransceiverController transceiverController)
+        {
+            _transceiverController = transceiverController;
+        }
         public static TciCwMacroSpeedUpCommand Create(ITransceiverController transceiverController)
         {
             Debug.Assert(transceiverController != null);
             return new TciCwMacroSpeedUpCommand(transceiverController);
         }
 
-        private TciCwMacroSpeedUpCommand(ITransceiverController transceiverController)
-        {
-            _transceiverController = transceiverController;
-        }
+
 
         public static string Name => "cw_macros_speed_up";
 
